improved coverage on samples and timeline
