+++ src/ambianic/webapp/server/samples.py
@@ -1,7 +1,7 @@
 """REST API for timeline events from pipeline samples."""
 import logging
 import uuid
-import datetime
+from datetime import datetime
 from pathlib import Path
 import os
 import json
@@ -156,7 +156,7 @@ def get_timeline(before_datetime=None, page=1, data_dir=None):
               'Sample index range [%d:%d]. ',
               page, page_size, page_start_position, page_end_position)
 
-    files = list(Path(data_dir).rglob('timeline-event-log.yaml*'))
+    files = list(Path(data_dir).rglob('./timeline-event-log.yaml*'))
     files = sorted(files, reverse=False)
 
     page_count = 1
